Minor fixes

+++ Client/SDL_project/PerlinNoise.h
@@ -28,7 +28,7 @@ public:
 	//! For gradient function
 	double grad(int hash, double x, double y, double z);
 
-	float getNoiseAmp() { return noiseAmp; }
+	double getNoiseAmp() { return noiseAmp; }
 private:
 	double noiseAmp = 180.0;
 };
